Cristian_Iacob_SCADA
This commit introduces a comprehensive set of files for simulating and monitoring SCADA system security. The project includes setup scripts for the attacker machine, SCADA server, and Wireshark monitor, along with tools for generating reports and visualizing data. The Flask-based web interface and SQLite database provide a user-friendly way to interact with and analyse SCADA events. Documentation is included to guide users through the simulation process.